+++ app/bg/ui/tabs/manager.js
@@ -1405,10 +1405,11 @@ function emitUpdateState (tab) {
   if (!tab.browserWindow) return
   var win = getTopWindow(tab.browserWindow)
   var index = typeof tab === 'number' ? tab : getAll(win).indexOf(tab)
-  if (index === -1) {
-    console.warn('WARNING: attempted to update state of a tab not on the window')
-    return
-  }
+  // Temporary fix
+  //if (index === -1) {
+    //console.warn('WARNING: attempted to update state of a tab not on the window')
+    //return
+  //}
   emit(win, 'update-state', {index, state: tab.state})
   win.emit('custom-pages-updated', takeSnapshot(win))
 }