+++ app/userland/desktop/css/main.css.js
@@ -113,6 +113,7 @@ header {
 
 .pin:hover {
   text-decoration: none;
+  top: -2px;
 }
 
 .pin .thumb {